requirements.txt
bert-serving-client==1.10.0
bert-serving-server==1.10.0
numpy==1.16.0
pandas==1.0.3
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-gpu==1.12.0
